requirements.txt
torch==2.4.1 --extra-index-url https://download.pytorch.org/whl/cu121
xformers==0.0.28 --extra-index-url https://download.pytorch.org/whl/cu121
flash-attn==2.6.1
git+https://github.com/MeiGen-AI/MultiTalk.git
moshi==0.2.7
gradio==4.27.0
soundfile==0.12.1
ffmpeg-python==0.2.0
google-generativeai==0.5.2
huggingface_hub==0.23.3
